Move some functions to commands/show
