+++ src/cli/commands/mod.rs
@@ -3,10 +3,12 @@ mod init;
 mod install;
 mod lock;
 mod pip;
+mod show;
 mod venv;
 pub use develop::develop;
 pub use init::init;
 pub use install::install;
 pub use lock::{bump_in_lock, metadata, update_lock};
 pub use pip::{get_frozen_deps, install_editable, install_editable_with_constraint, upgrade_pip};
+pub use show::{show_deps, show_outdated, show_venv_bin_path, show_venv_path};
 pub use venv::{clean_venv, create_venv, ensure_venv, expect_venv};
